Move every component in separate folder
