+++ src/components/PinInput/PinInput.js
