+++ src/components/QuickId/QuickId.css
