+++ src/components/QuickId/QuickId.js
