+++ src/components/QuickId/QuickId.vue
