+++ src/pages/Network.vue
@@ -27,7 +27,7 @@ import {
   AeButton,
   AeAmountInput
 } from '@aeternity/aepp-components'
-import QuickId from '@/components/QuickId.vue'
+import QuickId from '@/components/QuickId/QuickId.vue'
 
 export default {
   name: 'network',
