+++ src/pages/Unlock/Unlock.js
@@ -1,5 +1,5 @@
 import { AeButton } from '@aeternity/aepp-components'
-import PinInput from '@/components/PinInput.vue'
+import PinInput from '@/components/PinInput/PinInput.vue'
 import { PATHS } from '@/router/'
 
 export default {
